Lint et optimisations
