+++ lib/migration/migration-tools.js
@@ -2,9 +2,6 @@ import * as Ticket from '../models/ticket.js'
 import * as Subscription from '../models/subscription.js'
 import * as Membership from '../models/membership.js'
 import {formatPrice} from '../models/purchase.js'
-import {isAfter} from 'date-fns'
-import process from 'node:process'
-import util from 'node:util'
 
 /**
  * Calculates and returns the total quantity of each type of product in the provided array.
@@ -82,7 +79,7 @@ export function createPurchaseObjectTicket(product) {
     orderReference: null,
     productType,
     legacy: true,
-    price: formatPrice(product.tickets * ticketUnitPriceByDate(product.purchaseDate))
+    price: formatPrice(product.tickets * Ticket.ticketUnitPriceByDate(product.purchaseDate))
   }
 }
 
@@ -100,7 +97,7 @@ export function createPurchaseObjectMembership(product) {
     productType: 'membership',
     membershipStart: product.membershipStart,
     legacy: true,
-    price: formatPrice(membershipUnitPriceByDate(product.purchaseDate))
+    price: formatPrice(Membership.membershipUnitPriceByDate(product.purchaseDate))
   }
 }
 
@@ -118,50 +115,8 @@ export function createPurchaseObjectAbo(product) {
     productType: 'subscription',
     startDate: product.aboStart,
     legacy: true,
-    price: formatPrice(subscriptionUnitPriceByDate(product.purchaseDate))
+    price: formatPrice(Subscription.subscriptionUnitPriceByDate(product.purchaseDate))
   }
-}
-
-/**
- * Calculate the unit price of a subscription based on its date.
- *
- * @param {Date} date - The date the subscription was purchased.
- * @returns {number} The unit price of the subscription. If the date is after February 1, 2017, the price is 60; otherwise it's 0.
- */
-export function subscriptionUnitPriceByDate(date) {
-  if (isAfter(new Date(date), new Date('2017-02-01'))) {
-    return 60
-  }
-
-  return 0
-}
-
-/**
- * Calculate the unit price of a ticket based on its date.
- *
- * @param {Date} date - The date the ticket was purchased.
- * @returns {number} The unit price of the ticket. If the date is after February 1, 2017, the price is 60; otherwise it's 0.
- */
-export function ticketUnitPriceByDate(date) {
-  if (isAfter(new Date(date), new Date('2017-02-01'))) {
-    return 60
-  }
-
-  return 0
-}
-
-/**
- * Calculate the unit price of a membership based on its date.
- *
- * @param {Date} date - The date the membership was purchased.
- * @returns {number} The unit price of the membership. If the date is after January 1, 2014, the price is 10; otherwise it's 0.
- */
-export function membershipUnitPriceByDate(date) {
-  if (isAfter(new Date(date), new Date('2014-01-01'))) {
-    return 10
-  }
-
-  return 0
 }
 
 /**
