+++ lib/models/membership.js
@@ -1,5 +1,6 @@
 import mongo from '../util/mongo.js'
 import {customAlphabet} from 'nanoid'
+import {isAfter} from 'date-fns'
 
 const nanoid = customAlphabet('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz')
 
@@ -46,4 +47,18 @@ export async function updateMembership(_id, set) {
       $set: set
     }
   )
+}
+
+/**
+ * Calculate the unit price of a membership based on its date.
+ *
+ * @param {Date} date - The date the membership was purchased.
+ * @returns {number} The unit price of the membership. If the date is after January 1, 2014, the price is 10; otherwise it's 0.
+ */
+export function membershipUnitPriceByDate(date) {
+  if (isAfter(new Date(date), new Date('2014-01-01'))) {
+    return 10
+  }
+
+  return 0
 }
