+++ lib/models/subscription.js
@@ -72,3 +72,18 @@ export async function updateSubscription(_id, set) {
     }
   )
 }
+
+/**
+ * Calculate the unit price of a subscription based on its date.
+ *
+ * @param {Date} date - The date the subscription was purchased.
+ * @returns {number} The unit price of the subscription. If the date is after February 1, 2017, the price is 60; otherwise it's 0.
+ */
+export function subscriptionUnitPriceByDate(date) {
+  if (isAfter(new Date(date), new Date('2017-02-01'))) {
+    return 60
+  }
+
+  return 0
+}
+
