+++ lib/models/ticket.js
@@ -6,7 +6,7 @@ import {customAlphabet} from 'nanoid'
 
 const nanoid = customAlphabet('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz')
 
-// Const TICKET_UNIT_COST_IN_EUR = 6 // As of 2017-02-01
+const TICKET_UNIT_COST_IN_EUR = 6 // As of 2017-02-01
 
 export async function getMemberTickets(memberId) {
   const user = await mongo.db.collection('users')
@@ -17,9 +17,7 @@ export async function getMemberTickets(memberId) {
       _id: ticket.purchaseDate, // Until there is a unique id
       count: ticket.tickets,
       purchased: ticket.purchaseDate,
-      amount: isAfter(new Date(ticket.purchaseDate), new Date('2017-02-01'))
-        ? TICKET_UNIT_COST_IN_EUR * ticket.tickets
-        : 0
+      amount: ticketUnitPriceByDate(ticket.purchaseDate) * ticket.tickets
     }))
     .orderBy(['purchased'], ['desc'])
     .value()
@@ -95,4 +93,18 @@ export function computeTicketsQuantity(tickets, productType) {
   }
 
   return total
+}
+
+/**
+ * Calculate the unit price of a ticket based on its date.
+ *
+ * @param {Date} date - The date the ticket was purchased.
+ * @returns {number} The unit price of the ticket. If the date is after February 1, 2017, the price is 60; otherwise it's 0.
+ */
+export function ticketUnitPriceByDate(date) {
+  if (isAfter(new Date(date), new Date('2017-02-01'))) {
+    return TICKET_UNIT_COST_IN_EUR
+  }
+
+  return 0
 }
